got hall effect reading into degrees

+++ struct_and_randomization/Deck.cpp
@@ -1,7 +1,7 @@
 #include "Deck.h"
 
 Deck::Deck(short unsigned int q){
-  for(size_t i=1; i<266; i++){
+  for(short unsigned i=1; i<266; i++){
     card_list[i-1] = ((i*q)%265)+1;
   }
   card_pos = 0;
